requirements-dev.txt
# テスト用追加パッケージ
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-flask==1.3.0
pytest-timeout==2.1.0
coverage==7.3.2
